# docs/swagger.yaml
definitions:
...
  response.ApiResponseOrder:
    properties:
      data:
        $ref: '#/definitions/response.OrderResponse'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponsePaginationOrder:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationOrderDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationOrderItem:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderItemResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationOrderItemDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderItemResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ApiResponsesOrderItem:
    properties:
      data:
        items:
          $ref: '#/definitions/response.OrderItemResponse'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  response.OrderItemResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  response.OrderItemResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.OrderResponse:
    properties:
      cashier_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
    type: object
  response.OrderResponseDeleteAt:
    properties:
      cashier_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/order:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrder'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all orders
      tags:
      - Order
  /api/order-item:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all order items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of order items
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrderItem'
        "500":
          description: Failed to retrieve order item data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all order items
      tags:
      - Order-Item
  /api/order-item/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active order items
      produces:
      - application/json
      responses:
        "200":
          description: List of active order items
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrderItemDeleteAt'
        "500":
          description: Failed to retrieve order item data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active order items
      tags:
      - Order-Item
  /api/order-item/order/{order_id}:
    get:
      consumes:
      - application/json
      description: Retrieve order items by order ID
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of order items by order ID
          schema:
            $ref: '#/definitions/response.ApiResponsesOrderItem'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve order item data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find order items by order ID
      tags:
      - Order-Item
  /api/order-item/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed order items
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed order items
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrderItemDeleteAt'
        "500":
          description: Failed to retrieve order item data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed order items
      tags:
      - Order-Item
  /api/order/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order data
          schema:
            $ref: '#/definitions/response.ApiResponseOrder'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find order by ID
      tags:
      - Order
  /api/order/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active orders
      produces:
      - application/json
      responses:
        "200":
          description: List of active orders
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrderDeleteAt'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active orders
      tags:
      - Order
...
  /api/order/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed orders
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed orders
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationOrderDeleteAt'
        "500":
          description: Failed to retrieve order data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed orders
      tags:
      - Order
...
swagger: "2.0"